Check wether input and output dir exists
If the output directory does not exist target-gen will spend some time
generating the yml files just to panic in the end because it can't open
the output directory. If the input directory was mistyped you won't even
get an error, it just exits cleanly without doing anything -> Instead panic in
the beginning!